--- web/src/main/java/hk/onedegree/web/springboot/dev/UserRepository.java
@@ -1,8 +1,0 @@
-package hk.onedegree.web.springboot.dev;
-
-import org.springframework.data.repository.CrudRepository;
-
-
-public interface UserRepository extends CrudRepository<User, Integer> {
-
-}
